refactor(frontend): separate register pages components
