+++ frontend/src/components/index.js
@@ -3,5 +3,6 @@ export * from './Form'
 export * from './Invite'
 export * from './Navbar'
 export * from './PasswordValidator'
+export * from './Register'
 export * from './TextField'
 export * from './WithAuth'
